+++ packages/template-retail-react-app/config/default.js
@@ -32,8 +32,7 @@ module.exports = {
         login: {
             passwordless: {
                 enabled: false,
-                callbackURI:
-                    process.env.PASSWORDLESS_LOGIN_CALLBACK_URI || '/passwordless-login-callback',
+                mode: 'email',
                 landingPath: '/passwordless-login-landing'
             },
             social: {